backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
python-multipart
pydantic
groq
